Fixed ConvertKit Form
When I read the Monica Lent blog post the first time I misunderstood
that the form had an ID and a second ID that was used to populate the
tags. So I transposed variables in the form when I was trying to remove
YOUR from the variable names. Instead I just royally broke the form.
Going back to the orignal blog post and reviewing the example code
line by line revealed the error.